requirements.txt
requests==2.31.0
slack-sdk==3.23.0
openai>=1.50.0
schedule==1.2.0
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3 